feat: EmbedMermaidにSendWindowSizeを実装した
