+++ packages/zenn-embed-components/src/embed/mermaid/types.ts
@@ -1,0 +1,9 @@
+export type ErrorContainer = {
+  hasError: boolean;
+  message: string;
+};
+
+export type ValidateResult = {
+  charLimitOver: ErrorContainer;
+  chainingOfLinksOver: ErrorContainer;
+};
